chore: remove repeated methods after refactoring
