+++ packages/store/src/storage-utils.ts
@@ -98,24 +98,6 @@ export function cleanUpReadme(version: any): Version {
   return version;
 }
 
-export function normalizeContributors(contributors: Author[]): Author[] {
-  if (isNil(contributors)) {
-    return [];
-  } else if (contributors && _.isArray(contributors) === false) {
-    // FIXME: this branch is clearly no an array, still tsc complains
-    // @ts-ignore
-    return [contributors];
-  } else if (_.isString(contributors)) {
-    return [
-      {
-        name: contributors,
-      },
-    ];
-  }
-
-  return contributors;
-}
-
 export const WHITELIST = [
   '_rev',
   'name',
